Rebase/Revert switches in tests.
