+++ .gitignore
@@ -1,2 +1,7 @@
 *~
 target/
+
+# Eclipse specific files
+.settings
+.classpath
+.project
