+++ fcrepo-auth-roles-common/src/test/java/org/fcrepo/auth/roles/common/AccessRolesProviderTest.java
@@ -85,7 +85,7 @@ public class AccessRolesProviderTest {
         initMocks(this);
         when(node.getSession()).thenReturn(session);
 
-        provider = new AccessRolesProvider();
+        provider = new RbAclAccessRolesProvider();
 
         setupPrincipalNode(principalNode1, principalProperty1, "principal",
                 "role");
